// lib/dao/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
	{ [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
	{ [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
};
// ...
export type GroupMember = User | NonUser;

export type NonUser = {
	first_name?: Maybe<Scalars['String']>;
	last_name?: Maybe<Scalars['String']>;
	email?: Maybe<Scalars['String']>;
	is_user?: Maybe<Scalars['Boolean']>;
};

export type SendPicksResponse = {
	message?: Maybe<Scalars['String']>;
};
// ...
export type User = {
	userId: Scalars['ID'];
	first_name?: Maybe<Scalars['String']>;
	last_name?: Maybe<Scalars['String']>;
	email?: Maybe<Scalars['String']>;
	password?: Maybe<Scalars['String']>;
	groups?: Maybe<Array<Group>>;
	user_secret_picks?: Maybe<Array<SecretPick>>;
};

export type SecretPick = {
	group?: Maybe<Group>;
	group_member_info?: Maybe<GroupMember>;
	secret_pick?: Maybe<Scalars['String']>;
};

export type LoggedInUser = {
	userId?: Maybe<Scalars['String']>;
	token?: Maybe<Scalars['String']>;
};

export type AdditionalEntityFields = {
	path?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
};

import { ObjectID } from 'mongodb';
export type GroupDbObject = {
	_id?: Maybe<ObjectID>;
	title: string;
	members?: Maybe<Array<GroupMember>>;
};